requirements.txt
absl-py==1.0.0
affine==2.3.1
anyio==3.6.1
apipkg==2.1.1
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astroid==2.11.6
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
attrs==21.4.0
Automat==20.2.0
autopep8==1.6.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.10.0
black==22.6.0
bleach==5.0.0
blobfile==1.2.9
btrfsutil==5.18.1
CacheControl==0.12.6
cachetools==5.0.0
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
certifi==2022.6.15
cffi==1.15.0
chardet==4.0.0
ciso8601==2.2.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
clip==0.2.0
colorama==0.4.4
constantly==15.1.0
contextlib2==21.6.0
control==0.9.0
cryptography==37.0.2
cycler==0.11.0
DALL-E==0.1
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
deprecation==2.1.0
dill==0.3.4
distlib==0.3.4
distro==1.7.0
dnspython==2.2.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.3
filelock==3.6.0
flatbuffers==2.0
fonttools==4.28.3
ftfy==6.1.1
gast==0.5.3
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
gpg==1.17.1
graphviz==0.20
grpcio==1.46.0
h5py==3.6.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
imageio==2.13.3
importlib-metadata==4.8.1
incremental==21.3.0
iniconfig==1.1.1
ipykernel==6.15.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.1
isort==5.10.1
jaraco.context==4.1.1
jaraco.functools==3.5.0
jaraco.text==3.8.0
jedi==0.18.1
jeepney==0.7.1
Jinja2==3.1.2
joblib==1.1.0
json5==0.9.8
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.3
jupyter-core==4.10.0
jupyter-packaging==0.12.2
jupyter-server==1.18.0
jupyterlab==3.4.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.14.0
keras==2.8.0
Keras-Preprocessing==1.1.2
keyring==23.4.0
kiwisolver==1.3.2
lazy-object-proxy==1.7.1
lensfun==0.3.3
libclang==14.0.1
libmsym==0.2.4
llvmlite==0.38.0
lxml==4.8.0
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mediapipe==0.8.9.1
mistune==0.8.4
more-itertools==8.13.0
msgpack==1.0.3
mypy==0.942
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.6.4
nbconvert==6.5.0
nbformat==5.4.0
neat-python==0.92
nest-asyncio==1.5.4
netsnmp-python==1.0a1
networkx==2.6.3
notebook==6.4.12
notebook-shim==0.1.0
nspektr==0.4.0
numba==0.55.1
numpy==1.22.4
oauthlib==3.2.0
opencv-contrib-python==4.5.4.60
opencv-python==4.5.4.60
opt-einsum==3.3.0
ordered-set==4.0.2
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.2
pathspec==0.9.0
pep517==0.12.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.5.1
pluggy==1.0.0
ply==3.11
progress==1.6
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.19.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycodestyle==2.8.0
pycparser==2.21
pycryptodomex==3.14.1
Pygments==2.12.0
pylint==2.14.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
PyQt5==5.15.7
PyQt5-sip==12.11.0
pyrsistent==0.18.1
pyserial==3.5
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
PyWavelets==1.2.0
PyYAML==6.0
pyzmq==23.1.0
qtconsole==5.2.2
QtPy==2.1.0
rados==2.0.0
rasterio==1.3.2
rbd==2.0.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-unixsocket==0.3.0
resolvelib==0.5.5
retrying==1.3.3
rgw==2.0.0
rsa==4.8
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
scikit-image==0.19.0
scikit-learn==1.0.2
scipy==1.8.1
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
Shapely==1.8.2
sip==4.19.25
six==1.16.0
sklearn==0.0
slycot==0.4.0
sniffio==1.2.0
snuggs==1.4.7
soupsieve==2.3.1
stack-data==0.3.0
TBB==0.2
team==1.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.25.0
termcolor==1.1.0
terminado==0.15.0
testpath==0.6.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.0
torbrowser-launcher==0.3.5
torch==1.11.0
torchvision==0.12.0
tornado==6.1
tqdm==4.64.0
traitlets==5.3.0
trimesh==3.9.31
trove-classifiers==2022.6.23
tui-mmi-practical==0.11.1
Twisted==21.7.0
typing_extensions==4.2.0
urllib3==1.26.9
validate-pyproject==0.9
virtualenv==20.11.0
visvalingamwyatt==0.1.4
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
Werkzeug==2.1.2
widgetsnbextension==3.6.1
wrapt==1.14.1
xmltodict==0.12.0
yolk==0.4.3
zipp==3.8.0
zope.interface==5.4.0
